+++ .github/workflows/CI-orchestrator.yml
@@ -15,7 +15,12 @@ jobs:
     uses: ./.github/workflows/CI-test.yml
     with:
       CACHING_MODE: "CACHE_RESTORE"
-      
+
+  coverage-job:
+    uses: ./.github/workflows/CI-coverage.yml
+    with:
+      CACHING_MODE: "CACHE_RESTORE"
+
   lint-format-job:
     uses: ./.github/workflows/CI-lint-format.yml
     with:
@@ -75,6 +80,7 @@ jobs:
             OVERALL_STATUS="success"
             if [ "${{ needs.build-job.result }}" != "success" ] || 
                 [ "${{ needs.test-job.result }}" != "success" ] ||
+                [ "${{ needs.coverage-job.result }}" != "success" ] ||
                 [ "${{ needs.lint-format-job.result }}" != "success" ] || 
                 [ "${{ needs.zombienet-test-job.result }}" != "success" ]; then
               OVERALL_STATUS="failure"
@@ -107,6 +113,7 @@ jobs:
         run: |
           echo "Build Job Status: ${{ needs.build-job.result }}"
           echo "Test Job Status: ${{ needs.test-job.result }}"
+          echo "Coverage Job Status: ${{ needs.coverage-job.result }}"
           echo "Lint and Format Job Status: ${{ needs.lint-format-job.result }}"
           echo "Zombienet Test Job Status: ${{ needs.zombienet-test-job.result }}"
       - name: Download All Artifacts
@@ -187,7 +194,21 @@ jobs:
           else
             echo "TEST_JOB_EMOJI=:red_circle:" >> $GITHUB_ENV
           fi
-      
+
+          if [[ -z "${LINE_COVERAGE_PERCENT}" ]]; then
+            if [[ "${{ needs.coverage-job.result }}" == "success" ]]; then
+              echo "COVERAGE_JOB_EMOJI=:large_green_circle:" >> $GITHUB_ENV
+            else
+              echo "COVERAGE_JOB_EMOJI=:red_circle:" >> $GITHUB_ENV
+            fi
+          elif [[ "${{ needs.coverage-job.result }}" == "success" ]] && (( $(echo "$LINE_COVERAGE_PERCENT >= 70" | bc -l) )); then
+            echo "COVERAGE_JOB_EMOJI=:large_green_circle:" >> $GITHUB_ENV
+          elif [[ "${{ needs.coverage-job.result }}" == "success" ]]; then
+            echo "COVERAGE_JOB_EMOJI=:large_yellow_circle:" >> $GITHUB_ENV
+          else
+            echo "COVERAGE_JOB_EMOJI=:red_circle:" >> $GITHUB_ENV
+          fi
+
           if [[ "${{ needs.lint-format-job.result }}" == "success" ]]; then
             echo "LINT_FORMAT_JOB_EMOJI=:large_green_circle:" >> $GITHUB_ENV
           else
@@ -224,6 +245,13 @@ jobs:
                   "text": {
                     "type": "mrkdwn",
                     "text": "${{ env.TEST_JOB_EMOJI }} *Test:* ${{ needs.test-job.result }}\n\n${{ env.UNIT_TEST_SUMMARY }}\n${{ env.INTEGRATION_TEST_SUMMARY }}"
+                  }
+                },
+                {
+                  "type": "section",
+                  "text": {
+                    "type": "mrkdwn",
+                    "text": "${{ env.COVERAGE_JOB_EMOJI }} *Coverage:* ${{ needs.coverage-job.result }}\n\n${{ env.COVERAGE_SUMMARY }}"
                   }
                 },
                 {
