+++ runtime/src/tests/payout.rs
@@ -15,33 +15,153 @@
 
 //! This module tests the correct computation of rewards for validators.
 
+use crate::{payout::*, Balance, Perbill, Runtime, ACME, MILLICENTS};
 use pallet_staking::EraPayout;
 
-use crate::{Balance, Runtime};
-
-use super::test;
+#[test]
+fn check_params_sanity() {
+    // staking_target should be at least 0.5 (and so 0 <= s_c / s_t <= 2)
+    assert!(
+        StakingTarget::get() * 10u16 >= 5u16,
+        "too low staking target"
+    );
+    // base inflation is not too high
+    assert!(
+        InflationBase::get() * 100u64 == 5u64,
+        "unexpected base inflation"
+    );
+}
 
 #[test]
 fn check_era_rewards() {
-    test().execute_with(|| {
-        const ERA_DURATION_MILLIS: u64 = 6 * 60 * 60 * 1000;
-        const TOTAL_STAKED: Balance = 900000000;
-        const TOTAL_ISSUANCE: Balance = 1000000000;
+    const ERA_DURATION_MILLIS: u64 = 1000 * 60 * 60 * 24 * 36525 / 100; // 1 year era
+    const TOT_ISSUANCE: Balance = 1_000_000_000 * ACME;
+    let others_split = Percent::from_percent(100) - ValidatorsSplit::get();
 
-        // Check the reward for an empty era.
-        assert_eq!(
-            <Runtime as pallet_staking::Config>::EraPayout::era_payout(0, 0, ERA_DURATION_MILLIS),
-            (0u128, 0)
-        );
+    // Check the reward for an empty era.
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(0, 0, ERA_DURATION_MILLIS),
+        (0, 0)
+    );
 
-        // Check the reward for a normal era
-        assert_eq!(
-            <Runtime as pallet_staking::Config>::EraPayout::era_payout(
-                TOTAL_STAKED,
-                TOTAL_ISSUANCE,
-                ERA_DURATION_MILLIS
-            ),
-            (17313, 51133)
-        );
-    });
+    // Check the reward for a normal era, s_c == s_t
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(50) * tot_issuance;
+    let expected_inflation: u128 = 60000000 * ACME;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for a normal era, s_c == s_t / 2
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(25) * tot_issuance;
+    let expected_inflation: u128 = 66487212 * ACME + 70700 * MILLICENTS;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for a normal era, s_c == 0.0 (min)
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(0) * tot_issuance;
+    let expected_inflation: u128 = 77182818 * ACME + 28459 * MILLICENTS;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for a normal era, s_c == 1.0 (max)
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(100) * tot_issuance;
+    let expected_inflation: u128 = 53678794 * ACME + 41171 * MILLICENTS;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for an era with half the duration, s_c == 1.0 (max)
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(100) * tot_issuance;
+    let expected_inflation: u128 = (53678794 * ACME + 41171 * MILLICENTS) / 2;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS / 2
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for an era with double the duration, s_c == 1.0 (max)
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(100) * tot_issuance;
+    let expected_inflation: u128 = 53678794 * ACME + 41171 * MILLICENTS;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS * 2
+        ),
+        // capped at 1 year
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
+
+    // Check the reward for an era with zero duration, s_c == 1.0 (max)
+    let tot_issuance: Balance = TOT_ISSUANCE;
+    let tot_staked: Balance = Perbill::from_percent(100) * tot_issuance;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(tot_staked, tot_issuance, 0),
+        (0, 0)
+    );
+
+    // Check that tot_issuance is actually used, s_c == s_t
+    let tot_issuance: Balance = TOT_ISSUANCE / 2;
+    let tot_staked: Balance = Perbill::from_percent(50) * tot_issuance;
+    let expected_inflation: u128 = 30000000 * ACME;
+    assert_eq!(
+        <Runtime as pallet_staking::Config>::EraPayout::era_payout(
+            tot_staked,
+            tot_issuance,
+            ERA_DURATION_MILLIS
+        ),
+        (
+            ValidatorsSplit::get() * expected_inflation,
+            others_split * expected_inflation
+        )
+    );
 }
