+++ runtime/src/tests/use_correct_weights.rs
@@ -17,6 +17,8 @@
 
 use crate::{OnChainSeqPhragmen, Runtime};
 
+mod verifiers;
+
 #[test]
 fn frame_system() {
     use frame_system::WeightInfo;
@@ -232,100 +234,6 @@ fn pallet_proxy() {
     assert_eq!(
         <Runtime as pallet_proxy::Config>::WeightInfo::create_pure(1),
         crate::weights::pallet_proxy::ZKVWeight::<Runtime>::create_pure(1)
-    );
-}
-
-#[test]
-fn pallet_fflonk_verifier() {
-    use pallet_fflonk_verifier::Fflonk;
-    let dummy_proof = [0; pallet_fflonk_verifier::PROOF_SIZE];
-    let dummy_pubs = [0; pallet_fflonk_verifier::PUBS_SIZE];
-    use pallet_fflonk_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Fflonk>>::WeightInfo as pallet_verifiers::WeightInfo<Fflonk>>::submit_proof(
-            &dummy_proof,
-            &dummy_pubs
-        ),
-        crate::weights::pallet_fflonk_verifier::ZKVWeight::<Runtime>::submit_proof()
-    );
-}
-
-#[test]
-fn pallet_zksync_verifier() {
-    use pallet_zksync_verifier::Zksync;
-    let dummy_proof = [0; pallet_zksync_verifier::PROOF_SIZE];
-    let dummy_pubs = [0; pallet_zksync_verifier::PUBS_SIZE];
-    use pallet_zksync_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Zksync>>::WeightInfo as pallet_verifiers::WeightInfo<Zksync>>::submit_proof(
-            &dummy_proof,
-            &dummy_pubs
-        ),
-        crate::weights::pallet_zksync_verifier::ZKVWeight::<Runtime>::submit_proof()
-    );
-}
-
-#[test]
-fn pallet_groth16_verifier() {
-    use pallet_groth16_verifier::Groth16;
-    use pallet_groth16_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Groth16<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<Groth16<Runtime>>>
-            ::submit_proof(
-            &pallet_groth16_verifier::Proof::default(),
-            &Vec::new()
-        ),
-        crate::weights::pallet_groth16_verifier::ZKVWeight::<Runtime>::submit_proof_bn254(0)
-    );
-}
-
-#[test]
-fn pallet_settlement_risc0() {
-    use pallet_risc0_verifier::Risc0;
-    use pallet_risc0_verifier::WeightInfo;
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Risc0<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<Risc0<Runtime>>>
-            ::submit_proof(
-            &Vec::new(),
-            &Vec::new()
-        ),
-        crate::weights::pallet_risc0_verifier::ZKVWeight::<Runtime>::submit_proof_cycle_2_pow_13()
-    );
-}
-
-#[test]
-fn pallet_settlement_ultraplonk() {
-    use pallet_ultraplonk_verifier::{Ultraplonk, WeightInfo};
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<Ultraplonk<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<Ultraplonk<Runtime>>>
-            ::submit_proof(
-            &vec![0; pallet_ultraplonk_verifier::PROOF_SIZE],
-            &Vec::new()
-        ),
-        crate::weights::pallet_ultraplonk_verifier::ZKVWeight::<Runtime>::submit_proof_32()
-    );
-}
-
-#[test]
-fn pallet_settlement_proofofsql() {
-    use pallet_proofofsql_verifier::{ProofOfSql, WeightInfo};
-
-    assert_eq!(
-        <<Runtime as pallet_verifiers::Config<ProofOfSql<Runtime>>>::WeightInfo as
-            pallet_verifiers::WeightInfo<ProofOfSql<Runtime>>>
-            ::submit_proof(
-            &Vec::new(),
-            &Vec::new()
-        ),
-        crate::weights::pallet_proofofsql_verifier::ZKVWeight::<Runtime>::submit_proof()
     );
 }
